requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
click==7.1.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
Flask-Ext==0.1
Flask-HTTPAuth==4.0.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.11
mysqlclient==1.4.2.post1
protobuf==3.12.2
PyMySQL==0.9.3
pytz==2020.1
six==1.15.0
virtualenv==20.0.21
Werkzeug==1.0.1
